chore(component): move `src/components/Tables/TopPermittedQueries.tsx ` to `src/components/Tables/Domains/TopAllowedDomains.tsx`
  ## what
  - move `src/components/Tables/TopPermittedQueries.tsx ` to `src/components/Tables/Domains/TopAllowedDomains.tsx`
  - update references
    - ./src/pages/index.tsx

  ## how

  ## why
  - to organize components better

  ## where
  - ./src/components/Tables/TopPermittedQueries.tsx -> src/components/Tables/Domains/TopAllowedDomains.tsx
  - ./src/pages/index.tsx

  ## usage
